build: refactor server with encapsulation of server and database, ci and update docker setup and makefiles (#65)
## Ticket

**Title:** ETQ Dev, j’ai le server et la base de donnee securisé avec la
encapsulation
**Notion Link:**
https://www.notion.so/jkergal/ETQ-Dev-j-ai-le-server-et-la-base-de-donnee-securis-avec-la-encapsulation-972d5c9a14ec44919bae296c1f32b24f?pvs=4

## Description

- Encapsulation of `server` and `database`.
- Update `docker compose` and `DockerFile` to fit to our deploy.
- Update `ts-jest`.
- Update `Makefiles` to fit our docker setup.
- Update ci workflow to fit our docker setup.

## Guidelines

🔗 Link your PRs to their related Notion tickets
✍️ Use the commit convention to your pull request title
📃 If your code affects the build, update `README.md`
🚫 Do not merge a PR until all comments are resolved
🗑 Remove your branch after merging
