+++ client/Dockerfile
@@ -8,9 +8,31 @@ WORKDIR /app
 COPY package.json pnpm-lock.yaml ./
 RUN pnpm install
 
-FROM base AS build
+FROM dependencies AS build
 
 WORKDIR /app
 COPY . .
 COPY --from=dependencies /app/node_modules ./node_modules
+RUN pnpm build
+
+FROM dependencies AS development
+
+WORKDIR /app
+COPY --from=build /app/node_modules ./node_modules
 CMD pnpm start
+
+FROM dependencies AS test
+
+WORKDIR /app
+COPY --from=build /app/node_modules ./node_modules
+CMD pnpm test
+
+FROM build AS production
+# TODO: remove all dev dependencies and files, like src folder.
+WORKDIR /app
+RUN pnpm prune --prod
+COPY --from=build /app/build ./build
+COPY --from=build /app/node_modules ./node_modules
+RUN npm i -g serve
+
+CMD npx serve build
