+++ client/src/pages/signUp/index.ts
