+++ server/Makefile
@@ -1,4 +1,4 @@
-CONTAINER_NAME = server-codenaute
+CONTAINER_NAME = server-dev-codenaute
 
 # If the first argument is "run"...
 ifeq (exec,$(firstword $(MAKECMDGOALS)))
@@ -28,11 +28,11 @@ ci:
 
 start-debug: ## Starting server container in debug mode
 	@echo "Starting server in debug mode"
-	COMPOSE_DOCKER_CLI_BUILD=1 DOCKER_BUILDKIT=1 docker-compose up --build server
+	NODE_ENV=development COMPOSE_DOCKER_CLI_BUILD=1 DOCKER_BUILDKIT=1 docker-compose up --build server
 
 start-quiet: ## Starting server container in mode quiet
 	@echo "Starting server in quiet mode"
-	COMPOSE_DOCKER_CLI_BUILD=1 DOCKER_BUILDKIT=1 docker-compose up -d --build server
+	NODE_ENV=development COMPOSE_DOCKER_CLI_BUILD=1 DOCKER_BUILDKIT=1 docker-compose up -d --build server
 
 .PHONY: exec
 exec : ## Running server container
