+++ server/package.json
@@ -5,14 +5,14 @@
 	"main": "index.js",
 	"scripts": {
 		"build": "tsc",
+		"ci": "pnpm build && pnpm test",
 		"migration:generate": "typeorm-ts-node-esm migration:generate ./src/migrations/default -d ./src/db.ts",
 		"migration:run": "typeorm-ts-node-esm migration:run -d ./src/db.ts",
 		"start": "node dist/index.js",
 		"start:watch": "tsc -w & nodemon -L dist/index.js",
 		"test": "jest --runInBand --verbose --coverage",
 		"test:ci": "jest --runInBand --verbose",
-		"test:watch": "jest --watchAll --runInBand --verbose --coverage",
-		"ci": "pnpm build && pnpm test"
+		"test:watch": "jest --watchAll --runInBand --verbose --coverage"
 	},
 	"dependencies": {
 		"apollo-server": "3.11.1",
@@ -38,7 +38,7 @@
 		"@types/uuid": "^9.0.1",
 		"jest": "^29.5.0",
 		"nodemon": "^2.0.22",
-		"ts-jest": "^29.0.5",
+		"ts-jest": "^29.1.0",
 		"typescript": "~5.0",
 		"uuid": "^9.0.0"
 	},
