+++ server/src/resolvers/base/BaseResolver.ts
@@ -1,0 +1,4 @@
+import { Resolver } from 'type-graphql';
+
+@Resolver()
+export class BaseResolver {}
