+++ server/src/utils/enums/Environment.ts
@@ -1,0 +1,5 @@
+export enum Environment {
+	IS_DEVELOPMENT = 'development',
+	IS_PRODUCTION = 'production',
+	IS_TEST = 'test'
+}
