+++ server/tests/setup.jest.ts
@@ -1,4 +1,4 @@
-import { Database } from 'db';
+import { Database } from 'utils/configs/database';
 
 beforeAll(async () => {
 	jest.spyOn(console, 'info').mockImplementation(() => {});
